chore : renaming getTaskByTaskID folder into getTask
