+++ src/Task/repos/getTask/IGetTaskByTaskIDRepository.ts
@@ -1,4 +1,4 @@
-import { GetTaskDTO } from './../dtos/GetTaskDTO';
+import { GetTaskDTO } from '../dtos/GetTaskDTO';
 
 export interface GetTaskByTaskIDRepoInterface {
   getTaskByTaskID(taskID: string): Promise<GetTaskDTO>;
