+++ src/Task/repos/index.ts
@@ -1,5 +1,5 @@
 import { InMemoryCreateTaskRepository } from './createTask/impl/InMemoryCreateTaskRepository';
-import { InMemoryGetTaskByTaskIDRepository } from './getTaskByTaskID/impl/InMemoryGetTaskByTaskIDRepository';
+import { InMemoryGetTaskByTaskIDRepository } from './getTask/impl/InMemoryGetTaskByTaskIDRepository';
 import { InMemoryGetTasksRepository } from './getTasks/impl/InMemoryGetTasksRepository';
 
 const CreateTaskRepo = new InMemoryCreateTaskRepository();
